+++ backends/vulkan/runtime/graph/ops/Arithmetic.h
@@ -12,7 +12,7 @@
 
 #include <ATen/native/vulkan/impl/Arithmetic.h>
 
-#include <executorch/backends/vulkan/runtime/graph/Graph.h>
+#include <executorch/backends/vulkan/runtime/graph/ComputeGraph.h>
 
 namespace at {
 namespace native {
