+++ backends/vulkan/runtime/graph/ops/Staging.h
@@ -12,7 +12,7 @@
 
 #include <string.h>
 
-#include <executorch/backends/vulkan/runtime/graph/Graph.h>
+#include <executorch/backends/vulkan/runtime/graph/ComputeGraph.h>
 
 namespace at {
 namespace native {
